# app/Documentation/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Gerenciamento - BOB FLOW",
        "description": "API para gerenciamento BOB FLOW",
        "contact": {
            "email": "suporte@sistema.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/permissions": {
            "get": {
                "tags": [
                    "Permissões"
                ],
                "summary": "Listar todas as permissões",
                "operationId": "27bb1606a7b293d1d49f57bb94ae71f7",
                "responses": {
                    "200": {
                        "description": "Lista de permissões retornada com sucesso"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Permissões"
                ],
                "summary": "Criar uma nova permissão",
                "operationId": "37c918c86b5085b0eb088fa760c7175d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "editar_usuarios"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Permissão para editar usuários"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Permissão criada com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    },
                    "500": {
                        "description": "Erro interno no servidor"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Entre com o token no formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}